% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBuoy.R
\name{fetch_buoy_data}
\alias{fetch_buoy_data}
\title{Fetch Muskegon Lake Buoy Data}
\usage{
fetch_buoy_data(
  x = "date",
  y,
  start_date = "10/01/2022",
  end_date = "10/31/2022",
  time = NULL,
  calculation = NULL
)
}
\arguments{
\item{x}{A single x value may be used.
If no value is specified, date will be used as the default.}

\item{y}{At least one y value is required.
Multiple y values should be separated by a comma.}

\item{start_date}{A specific date in time that marks the beginning of a particular period.
Date values must be in the format MM/DD/YYYY or M/D/YY.
The default value is 10/01/2022.}

\item{end_date}{A specific date in time that marks the end of a particular period.
Values must be in the format MM/DD/YYYY or M/D/YY.
The default date is 10/31/2022.}

\item{time}{Time values must be in the 24-hour format H:MM or HH:MM.
The values must also be divisible by 15 minutes (0:00, 0:15, 0:30, and 0:45).
If no value is provided then all times are considered.}

\item{calculation}{The mathematical calculation performed on the range points within an x variable's concentration.}
}
\value{
A data frame based on the provided parameters.
}
\description{
This function fetches data on current and historical conditions of Muskegon Lake in Muskegon County, Michigan from the Muskegon Lake Buoy API.
Grand Valley State University's Robert B. Annis Water Resources Institute (AWRI) established the buoy-based observatory in 2010. More information on the Muskegon Lake Buoy is available on the AWRI website: \url{https://www.gvsu.edu/wri/buoy/}.
The API documentation is available at \url{https://www.gvsu.edu/wri/buoy/data-api.htm}.
}
\details{
The arguments for x and y must be supported values. More information can be found at Grand Valley State University Muskegon Lake Buoy API website: \url{https://www.gvsu.edu/wri/buoy/data-api.htm}
}
\examples{
fetch_buoy_data(x="weekday",y="atmp1,tp001,tp002", start_date="7/7/11", end_date="7/28/11")
}
\author{
Beatrice Ngigi

Prof. Andrew DiLernia
}
