% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBuoy.R
\name{muskegonLakeBuoyData}
\alias{muskegonLakeBuoyData}
\title{Extracting Muskegon Lake Buoy Data}
\usage{
muskegonLakeBuoyData(
  y,
  x = NULL,
  date = NULL,
  time = NULL,
  concentration = NULL,
  calculation = NULL
)
}
\arguments{
\item{y}{At least one y value is required.
Multiple y values should be separated by a comma.}

\item{x}{A single x value may be used.}

\item{date}{Date values must be in the format MM/DD/YYYY or M/D/YY and can be either ranges separated by hyphens or commas.
If no value is provided then all dates are considered.}

\item{time}{Time values must be in the 24-hour format H:MM or HH:MM and can be either ranges separated by hyphens or commas.
The values must also be divisible by 15 minutes (0:00, 0:15, 0:30, and 0:45).
If no value is provided then all times are considered.}

\item{concentration}{A decimal number in the range .01-1.0 which represents the total amount of y plots per x variable.
The closer to 1 this number is, the more points will be present in the output and the more accurate the graph.
If no value is provided then all points are considered.}

\item{calculation}{The mathematical calculation performed on the range points within an x variable's concentration.}
}
\value{
A data frame based on the provided parameters.
}
\description{
This function obtains Muskegon Lake Buoy Data through Application Programming Interface.
The data gives the current conditions and historic data from the Muskegon Lake buoy since 2011.
}
\note{
The parameters must use supported values. Refer to <a href="https://www.gvsu.edu/wri/buoy/data-api.htm">GVSU Muskegon Lake Buoy website</a> for more information.
}
\author{
* Beatrice Ngigi
* Andrew DiLernia
}
