% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBuoy.R
\name{plot_buoy_data}
\alias{plot_buoy_data}
\title{Plot Muskegon Lake Buoy Data}
\usage{
plot_buoy_data(x = "date", y = NULL, filter = NULL, graph_type)
}
\arguments{
\item{x}{The values to be plotted on the x-axis of the graph. If no value is specified, the date will be used as the default.}

\item{y}{The values to be used on the y-axis of the graph.
Multiple y values should be separated by a comma.}

\item{graph_type}{The type of chart to be plotted.
The supported types are scatter, line, bar, and boxplot.}
}
\value{
A graph of x values by y values
}
\description{
This function offers a simple method to retrieve and display data from the Muskegon Lake Buoy through the use of Application Programming Interface.
The raw data is transformed into an easy-to-read and visually appealing graph, facilitating comprehension and analysis.
}
\details{
The arguments for x and y must be supported values. More information can be found at Grand Valley State University Muskegon Lake Buoy website: \url{https://www.gvsu.edu/wri/buoy/data-api.htm}
}
\examples{
plot_buoy_data(x="weekday",y="rh1", graph_type="line")
}
\author{
Beatrice Ngigi

Andrew DiLernia
}
