% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBuoy.R
\name{muskegonLakeBuoyData}
\alias{muskegonLakeBuoyData}
\title{Extracting Muskegon Lake Buoy Data}
\usage{
muskegonLakeBuoyData(
  y,
  x = NULL,
  date = NULL,
  time = NULL,
  calculation = NULL,
  concentration = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{y}{At least one y value is required.
Multiple y values should be separated by a comma.}

\item{x}{A single x value may be used.}

\item{date}{Date values must be in the format MM/DD/YYYY or M/D/YY.
If no value is provided then all dates are considered.}

\item{time}{Time values must be in the 24-hour format H:MM or HH:MM.
The values must also be divisible by 15 minutes (0:00, 0:15, 0:30, and 0:45).
If no value is provided then all times are considered.}

\item{calculation}{The mathematical calculation performed on the range points within an x variable's concentration.}

\item{concentration}{A decimal number in the range .01-1.0 which represents the total amount of y plots per x variable.
The closer to 1 this number is, the more points will be present in the output and the more accurate the graph.
If no value is provided then all points are considered.}
}
\value{
A data frame based on the provided parameters.
}
\description{
This function is a handy tool that allows the user to retrieve valuable information about the current conditions and historical data of the Muskegon Lake buoy, all through the convenience of an Application Programming Interface.
It provides comprehensive understanding of the lake's conditions over time as it grants access to a wealth of data dating back to 2011.
}
\details{
The arguments must use supported values. More information can be found at Grand Valley State University Muskegon Lake Buoy website: \url{https://www.gvsu.edu/wri/buoy/data-api.htm}
}
\examples{
muskegonLakeBuoyData(y="atmp1,tp001,tp002", x="date", date="7/7/11,7/14/11,7/21/11,7/28/11", concentration="1")
}
\author{
Beatrice Ngigi

Andrew DiLernia
}
