% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{muskegonLakeBuoyData}
\alias{muskegonLakeBuoyData}
\title{Muskegon Lake Buoy Data}
\usage{
muskegonLakeBuoyData(
  x = NULL,
  y,
  date = NULL,
  time = NULL,
  concentration = NULL,
  calculation = NULL
)
}
\description{
Extracts the Muskegon Lake Buoy Data
}
